# app/schemas/donation.py
from datetime import datetime
from typing import Optional

from pydantic import BaseModel, Extra, PositiveInt


class DonationCreate(BaseModel):
    comment: Optional[str]
    full_amount: PositiveInt

    class Config:
        extra = Extra.forbid


class UserDonation(DonationCreate):
    id: int
    create_date: datetime

    class Config:
        orm_mode = True


class DonationDB(UserDonation):
    user_id: int
    invested_amount: int
    fully_invested: bool
    close_date: Optional[datetime]